2nd reoganisation step: Clean the files (remove duplicated code)
At this stage of the reorganisation the idea is to remove the code that
should not belong to each module.

Since no other change is introduced, the modules are left unusable,
because of broken/missing imports. That will be addressed in the
following commit.
